Move index.ts - part 2
