+++ benchmark/index.ts
@@ -1,7 +1,7 @@
 import { execSync } from 'child_process';
 import { create_csv_report } from './csv';
 import { writeFileSync } from 'fs';
-import { nat64 } from '../index';
+import { nat64 } from '../src/';
 import { create_markdown_report } from './markdown';
 import { run_setup } from './setup';
