+++ src/index.ts
@@ -1,9 +1,9 @@
 import { Principal } from '@dfinity/principal';
-import { RequireExactlyOne, Variant } from './variant';
+import { RequireExactlyOne, Variant } from '../variant';
 
-export { StableBTreeMap, InsertError } from './src/stable_b_tree_map';
-export { Variant } from './variant';
-export { match } from './variant';
+export { StableBTreeMap, InsertError } from './stable_b_tree_map';
+export { Variant } from '../variant';
+export { match } from '../variant';
 
 declare var globalThis: any;
