+++ test/index.ts
@@ -1,4 +1,4 @@
-import { Variant } from '../index';
+import { Variant } from '../src';
 import { execSync } from 'child_process';
 
 export type Test = {
